src/GUI/requirements.txt
PyQt6>=6.4
pyyaml>=6.0
